Backend/requirements.txt
django-extensions==3.2.3
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
